+++ src/titiler/application/tests/routes/test_stac.py
@@ -20,7 +20,7 @@ def test_bounds(httpx, app):
     assert len(body["bounds"]) == 4
 
 
-@patch("rio_tiler.io.cogeo.rasterio")
+@patch("rio_tiler.io.rasterio.rasterio")
 @patch("rio_tiler.io.stac.httpx")
 def test_info(httpx, rio, app):
     """test /info endpoint."""
@@ -67,7 +67,7 @@ def parse_img(content: bytes) -> Dict:
             return dst.meta
 
 
-@patch("rio_tiler.io.cogeo.rasterio")
+@patch("rio_tiler.io.rasterio.rasterio")
 @patch("rio_tiler.io.stac.httpx")
 def test_tile(httpx, rio, app):
     """test tile endpoints."""
@@ -88,7 +88,7 @@ def test_tile(httpx, rio, app):
     assert meta["height"] == 256
 
     response = app.get(
-        "/stac/tiles/9/289/207?url=https://myurl.com/item.json&expression=B01&rescale=0,1000"
+        "/stac/tiles/9/289/207?url=https://myurl.com/item.json&expression=B01_b1&rescale=0,1000"
     )
     assert response.status_code == 200
     assert response.headers["content-type"] == "image/png"
@@ -97,7 +97,7 @@ def test_tile(httpx, rio, app):
     assert meta["height"] == 256
 
 
-@patch("rio_tiler.io.cogeo.rasterio")
+@patch("rio_tiler.io.rasterio.rasterio")
 @patch("rio_tiler.io.stac.httpx")
 def test_tilejson(httpx, rio, app):
     """test /tilejson endpoint."""
@@ -134,7 +134,7 @@ def test_tilejson(httpx, rio, app):
     )
 
 
-@patch("rio_tiler.io.cogeo.rasterio")
+@patch("rio_tiler.io.rasterio.rasterio")
 @patch("rio_tiler.io.stac.httpx")
 def test_preview(httpx, rio, app):
     """test preview endpoints."""
@@ -164,7 +164,7 @@ def test_preview(httpx, rio, app):
     assert meta["height"] == 128
 
     response = app.get(
-        "/stac/preview?url=https://myurl.com/item.json&expression=B01&rescale=0,1000&max_size=64"
+        "/stac/preview?url=https://myurl.com/item.json&expression=B01_b1&rescale=0,1000&max_size=64"
     )
     assert response.status_code == 200
     assert response.headers["content-type"] == "image/png"
@@ -173,7 +173,7 @@ def test_preview(httpx, rio, app):
     assert meta["height"] == 64
 
 
-@patch("rio_tiler.io.cogeo.rasterio")
+@patch("rio_tiler.io.rasterio.rasterio")
 @patch("rio_tiler.io.stac.httpx")
 def test_part(httpx, rio, app):
     """test crop endpoints."""
@@ -205,7 +205,7 @@ def test_part(httpx, rio, app):
     assert meta["height"] == 128
 
     response = app.get(
-        "/stac/crop/23.878,32.063,23.966,32.145.png?url=https://myurl.com/item.json&expression=B01&rescale=0,1000&max_size=64"
+        "/stac/crop/23.878,32.063,23.966,32.145.png?url=https://myurl.com/item.json&expression=B01_b1&rescale=0,1000&max_size=64"
     )
     assert response.status_code == 200
     assert response.headers["content-type"] == "image/png"
@@ -214,7 +214,7 @@ def test_part(httpx, rio, app):
     assert meta["height"] == 14
 
 
-@patch("rio_tiler.io.cogeo.rasterio")
+@patch("rio_tiler.io.rasterio.rasterio")
 @patch("rio_tiler.io.stac.httpx")
 def test_point(httpx, rio, app):
     """test crop endpoints."""
@@ -231,26 +231,28 @@ def test_point(httpx, rio, app):
     assert response.status_code == 200
     body = response.json()
     assert body["coordinates"] == [23.878, 32.063]
-    assert body["values"] == [[3565]]
+    assert body["values"] == [3565]
+    assert body["band_names"] == ["B01_b1"]
 
     response = app.get(
-        "/stac/point/23.878,32.063?url=https://myurl.com/item.json&assets=B01&asset_expression=B01|b1*2"
+        "/stac/point/23.878,32.063?url=https://myurl.com/item.json&expression=B01_b1*2"
     )
     assert response.status_code == 200
     body = response.json()
     assert body["coordinates"] == [23.878, 32.063]
-    assert body["values"] == [[7130]]
+    assert body["values"] == [7130]
+    assert body["band_names"] == ["B01_b1*2"]
 
     response = app.get(
-        "/stac/point/23.878,32.063?url=https://myurl.com/item.json&expression=B01/B09"
+        "/stac/point/23.878,32.063?url=https://myurl.com/item.json&expression=B01_b1/B09_b1"
     )
     assert response.status_code == 200
     body = response.json()
     assert body["coordinates"] == [23.878, 32.063]
-    assert round(body["values"][0][0], 2) == 0.49
+    assert round(body["values"][0], 2) == 0.49
 
 
-@patch("rio_tiler.io.cogeo.rasterio")
+@patch("rio_tiler.io.rasterio.rasterio")
 @patch("rio_tiler.io.stac.httpx")
 def test_missing_asset_not_found(httpx, rio, app):
     """test /info endpoint."""