+++ CHANGES.md
@@ -5,12 +5,18 @@
 ### titiler.core
 
 * update FastAPI requirement to `>=0.87`
+* update rio-tiler requirement to `>=4.0,<4.1`
 
 **breaking changes**
 
 * remove `timing headers` and `titiler.core.utils` submodule
+* remove `asset_expression` (except in `/asset_statistics` endpoint) (see https://cogeotiff.github.io/rio-tiler/v4_migration/#multibasereader-expressions)
+* update Point output model to include `band_names`
+* histogram and info band names are prefixed with `b` (e.g `b1`) (ref: https://cogeotiff.github.io/rio-tiler/v4_migration/#band-names)
 
 ### titiler.mosaic
+
+* update cogeo-mosaic requirement to `>=4.2,<4.3`
 
 **breaking changes**